% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpkm_calc.R
\name{fpkm_calc}
\alias{fpkm_calc}
\title{Calculate FPKM}
\usage{
fpkm_calc(read_counts, gene_lengths, total_reads)
}
\arguments{
\item{read_counts}{A numeric vector of read counts per gene}

\item{gene_lengths}{A numeric vector of gene lengths in base pairs}

\item{total_reads}{A numeric value representing the total number of mapped reads in the RNA-seq experiment}
}
\value{
A numeric vector of FPKM values, named by gene
}
\description{
This function calculates the FPKM (Fragments Per Kilobase of transcript per Million mapped reads)
for each gene in RNA-seq data
}
